pyproject.toml
[project]
name = "ysirsky-music-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mutagen>=1.47.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "ruff>=0.11.9",
    "spotdl>=4.2.11",
    "tinytag>=2.1.1",
]
